+++ src/main/java/tn/test/spring/Services/Produit/ProduitService.java
@@ -4,6 +4,6 @@ import tn.test.spring.Entity.Produit;
 import tn.test.spring.Services.GService;
 
 public interface ProduitService extends GService<Produit, Long> {
-    void assignProduitToStock(Long idProduit, Long idStock);
+    void assignProduitToStock(Long idProduit, int idStock);
 
 }
